Miscellaneous fixes to JAX backend:     - Enable reshape_test and onehot_categorical_test.     - Add softmax_cross_entropy_with_logits, and bfloat16 in JAX mode.     - Cast log(pi) in LKJ to avoid dtype issues.     - Add name property to NumpyVariable. This fixes errors with concretization tests (which try to display the name of the Variable).
PiperOrigin-RevId: 293285359

+++ tensorflow_probability/python/distributions/lkj.py
@@ -417,7 +417,7 @@ class LKJ(distribution.Distribution):
       concentration = (
           tf.convert_to_tensor(self.concentration
                                if concentration is None else concentration))
-      logpi = np.log(np.pi)
+      logpi = float(np.log(np.pi))
       ans = tf.zeros_like(concentration)
       for k in range(1, self.dimension):
         ans = ans + logpi * (k / 2.)
